test(richtext-editor): migrate Enter key integration tests from state to editor tests
