+++ richtext-editor/src/test/kotlin/dev/mkeeda/arranger/richtext/editor/RichTextStateTest.kt
@@ -4,8 +4,6 @@ import androidx.compose.ui.text.TextRange
 import dev.mkeeda.arranger.richtext.BlockquoteKey
 import dev.mkeeda.arranger.richtext.BoldKey
 import dev.mkeeda.arranger.richtext.BulletListKey
-import dev.mkeeda.arranger.richtext.HeadingKey
-import dev.mkeeda.arranger.richtext.HeadingLevel
 import dev.mkeeda.arranger.richtext.ItalicKey
 import dev.mkeeda.arranger.richtext.ListIndentLevel
 import dev.mkeeda.arranger.richtext.RgbaColor
@@ -569,87 +567,6 @@ class RichTextStateTest {
         val spans = state.richString.spans
         spans.size shouldBe 1
         spans.first().range shouldBe expectedText.rangeOf("Bullet item")
-        spans.first().attributes shouldBe attributeContainerOf(BulletListKey to ListIndentLevel.Level1)
-    }
-
-    @Test
-    fun `typing newline in heading clears heading attribute for the new paragraph`() {
-        val initialText = "Heading"
-        val state =
-            RichTextState(
-                initialText =
-                    RichString(initialText).edit {
-                        setParagraphAttribute(HeadingKey, HeadingLevel.H1, initialText.indices)
-                    },
-            )
-
-        state.textFieldState.edit {
-            selection = TextRange(initialText.length)
-            replace(length, length, "\n")
-            state.updateRichString(this)
-        }
-
-        val expectedText = "Heading\n"
-        state.richString.text shouldBe expectedText
-
-        val spans = state.richString.spans
-        spans.size shouldBe 1
-        spans.first().range shouldBe expectedText.rangeOf("Heading\n")
-        spans.first().attributes shouldBe attributeContainerOf(HeadingKey to HeadingLevel.H1)
-    }
-
-    @Test
-    fun `typing newline in an empty list item outdents the list level`() {
-        val initialText = "List\n"
-        val state =
-            RichTextState(
-                initialText =
-                    RichString(initialText).edit {
-                        setParagraphAttribute(BulletListKey, ListIndentLevel.Level2, 0..initialText.length)
-                    },
-            )
-
-        state.textFieldState.edit {
-            selection = TextRange(initialText.length)
-            replace(length, length, "\n")
-            state.updateRichString(this)
-        }
-
-        // The newline should be consumed by the outdent operation
-        val expectedText = "List\n"
-        state.richString.text shouldBe expectedText
-
-        val spans = state.richString.spans
-        spans.size shouldBe 2
-        spans[0].range shouldBe expectedText.rangeOf("List\n")
-        spans[0].attributes shouldBe attributeContainerOf(BulletListKey to ListIndentLevel.Level2)
-        spans[1].range shouldBe (5..5)
-        spans[1].attributes shouldBe attributeContainerOf(BulletListKey to ListIndentLevel.Level1)
-    }
-
-    @Test
-    fun `typing newline in an empty level 1 list item clears the list attribute`() {
-        val initialText = "List\n"
-        val state =
-            RichTextState(
-                initialText =
-                    RichString(initialText).edit {
-                        setParagraphAttribute(BulletListKey, ListIndentLevel.Level1, 0..initialText.length)
-                    },
-            )
-
-        state.textFieldState.edit {
-            selection = TextRange(initialText.length)
-            replace(length, length, "\n")
-            state.updateRichString(this)
-        }
-
-        val expectedText = "List\n"
-        state.richString.text shouldBe expectedText
-
-        val spans = state.richString.spans
-        spans.size shouldBe 1
-        spans.first().range shouldBe (0..4)
         spans.first().attributes shouldBe attributeContainerOf(BulletListKey to ListIndentLevel.Level1)
     }
 }
